Update self-hosted Spaces docs to reflect profile changes
Users now need a profile to interact with Spaces, even when they're
self-hosted. Update the self-hosted Spaces setup docs to include creating a
profile and update the `up ctx` commands to work correctly with the new
profiles. Update the local quick-start to work with the profile created by `up
space init`.